@@ -1,16 +1,9 @@
-module.exports = {
-  purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
-  darkMode: false, // or 'media' or 'class'
-  theme: {
-    extend: {
-      colors: {
-        primary: '#4CAF50',
-        secondary: '#FF5722',
-      },
+/** @type {import('tailwindcss').Config} */
+export default {
+    content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
+    darkMode: "class",
+    theme: {
+        extend: {},
     },
-  },
-  variants: {
-    extend: {},
-  },
-  plugins: [],
+    plugins: [],
 };
